refactor(chat): move chat to a separate lib crate
