+++ Cargo.toml
@@ -50,13 +50,15 @@ clap = { version = "4.5.32", features = [
     "wrap_help",
 ] }
 cocoa = "0.26.0"
+color-print = "0.3.5"
 convert_case = "0.8.0"
 core-foundation = "0.10.0"
 core-foundation-sys = "0.8.7"
 core-graphics = "0.24.0"
-crossterm = "0.28.1"
+crossterm = { version = "0.28.1", features = ["event-stream", "events"] }
 dashmap = "6.0.1"
 dirs = "5.0.0"
+eyre = "0.6.8"
 fig_api_client = { path = "crates/fig_api_client" }
 fig_auth = { path = "crates/fig_auth" }
 fig_aws_common = { path = "crates/fig_aws_common" }
@@ -104,6 +106,7 @@ objc2-input-method-kit = "0.2.2"
 parking_lot = "0.12.3"
 percent-encoding = "2.2.0"
 portable-pty = "0.8.1"
+q_chat = { path = "crates/q_chat" }
 r2d2 = "0.8.10"
 r2d2_sqlite = "0.25.0"
 rand = "0.9.0"
@@ -130,6 +133,7 @@ serde_json = "1.0.140"
 sha2 = "0.10.6"
 shlex = "1.3.0"
 similar = "2.7.0"
+spinners = "4.1.0"
 strum = { version = "0.27.1", features = ["derive"] }
 sysinfo = "0.32.0"
 thiserror = "2.0.12"
