+++ crates/q_chat/src/command.rs
