+++ crates/q_chat/src/consts.rs
