+++ crates/q_chat/src/context.rs
