+++ crates/q_chat/src/conversation_state.rs
