+++ crates/q_chat/src/input_source.rs
