+++ crates/q_chat/src/message.rs
