+++ crates/q_chat/src/parse.rs
