+++ crates/q_chat/src/parser.rs
