+++ crates/q_chat/src/prompt.rs
