+++ crates/q_chat/src/shared_writer.rs
