+++ crates/q_chat/src/skim_integration.rs
