+++ crates/q_chat/src/tools/execute_bash.rs
