+++ crates/q_chat/src/tools/fs_read.rs
