+++ crates/q_chat/src/tools/fs_write.rs
