+++ crates/q_chat/src/tools/mod.rs
