+++ crates/q_chat/src/tools/tool_index.json
